requirements.txt
emoji==0.5.1
Django==2.1.7
python-telegram-bot==11.1.0
